fix(audit): /code-audit pr 546 re-invocation — partition reviewer findings
Fresh /code-audit invocation surfaced 10 HIGH + 24 MEDIUM + 41 LOW + 41 NIT.
P2/P7 partitions had zero HIGH; remaining 5 partitions had fixers. Highlights:

- file-size cap closures: CMakeLists.txt extracted CPack to
  cmake/PhosphorCPack.cmake (800→774), pipewireconnection.cpp split
  lifecycle code into new pipewireconnection_lifecycle.cpp (869→621),
  test_phosphor_registry_pluginloader.cpp extracted lifecycle slots
  into a sibling TU + shared helpers header (854→410).
- pwnodemodel: HIGH emit-order correctness — connectionChanged and
  mediaClassesChanged now emit AFTER rebuildFromConnection (was
  before), matching the documented invariant ("rows always describe
  current connection property"). Slot/binding awakened by the signal
  now sees consistent connection+rows. Pass-2 fix from prior invocation
  was reverted as it broke the invariant.
- qmlregistration: process-static PipeWireHost now parented to
  QCoreApplication::instance() so it tears down during Qt's child-
  destruction phase rather than at static cleanup after QGuiApplication
  is gone.
- pipewirehost: explicit dtor now calls disconnectFromDaemon for
  ctor/dtor symmetry.
- pwnodemodel: nodeAdded handler races between connect() and snapshot
  closed via rowIndex.contains(node) guard; subclass constructors
  bypass setMediaClasses signal-from-ctor via direct Private friend
  assignment; nodeWires.contains assert added; rowIndex shift comment
  added.
- pipewireconnection: stale doc references swept (README, docs, CMake
  comments); cross-thread invariants block consolidated at file top;
  isConnected/isDaemonAvailable use acquire ordering for consistency;
  resetGuiSnapshot order reversed (strings first then booleans so
  observers chained off connectedChanged see cleared default-sink/
  source names); isAudioNodeClass uses size+memcmp for embedded-NUL
  safety; oversized metadata warning logs actual length; pw_context_new
  + pw_context_connect failure paths defensively reset pendingSyncSeq.
- palettestore: validateDoc consolidated into extractValidTokens
  (single QColor-parse pass cached); loadFromJson restructured for
  true atomic-on-failure (watcher drop AFTER applyPalette success);
  dropWatcherAndClearSourcePath dead isEmpty guards removed;
  redundant incomingCanonical round-trip dropped; isReload parameter
  dropped from readParseAndApply (YAGNI); fileChanged raw-equality
  documented as safe via canonical round-trip; trailing return false
  replaced with exhaustive switch.
- pluginloader: stale scheduleUnload comment fixed; m_registry non-
  null invariant pulled to single class-level comment; loadPluginFromDir
  factory wrapped into shared_ptr on a single statement for exception
  safety.
- CLI: usage banner <name> → <target>; volume echo epsilon
  env-overridable via PHOSPHOR_PW_VOLUME_ECHO_EPSILON; cmdMute comment
  rewritten; conn1 → qtConn rename for distinguishability;
  EX_SOFTWARE rc=70 mentioned in diagnostic.
- tests: stale "parseAndApplyJson(false) fix" comment fixed; QWARN()
  instead of qWarning() for proper test tagging; smoke test docstrings
  trimmed; hotReload wait extended 2s→5s for slow CI.
- shell.qml: clockText defensive floor against pre-update SystemClock
  state; togglePopup no-op default closure; cpu/memInfo string typing
  documented; locale-aware date format; shellState alias shape
  documented; batteryAvailable hoisted; NaN-safe Math.round guard;
  hostname-once comment added.
- PopoutHost: backdrop visible bound to cycle state not opacity-step;
  rebindContentItem guards Loader-detach; Loader gets anchors.fill;
  MouseArea gets Accessible.name; click routes through dismiss()
  not direct property set; contentItem-null orphan warn.
- build: new cmake/PhosphorLibTesting.cmake helper consolidates the
  standalone-or-tests gate across phosphor-* libs; cmake/PhosphorCPack
  .cmake extracts top-level CPack source-tarball; engine QML firewall
  driven by PHOSPHOR_ENGINE_TARGETS variable; fixture-manifest abi
  placeholder check added.
