+++ docs/phosphor-shell-design/04-implementation-plan.md
@@ -262,14 +262,14 @@ Naming notes:
 
 **Phase 2 gate:** each library has a passing CLI demo; `phosphor-services` umbrella is deleted; `phosphorctl call <ns>.<fn>` works for every service. Tag `phosphor-integrations-0.1`.
 
-### 2.1: `phosphor-service-pipewire` post-mortem *(shipped on feat/phase-2.1-pipewire-service)*
+### 2.1: `phosphor-service-pipewire` post-mortem *(shipped)*
 
 This section is the post-implementation record of the shape that landed on `feat/phase-2.1-pipewire-service`. The library shipped end-to-end (milestones 1-10), so the milestone list below documents what is in the tree, not a forward pickup plan. Where the shipped surface diverges from the original sketch in `02-gap-analysis.md`, this section is authoritative.
 
 **Milestones**
 
 1. **Skeleton + CMake plumbing (S, ~1 day).** `libs/phosphor-service-pipewire/` directory with the same shape Phase 2.0 established: `CMakeLists.txt`, `PhosphorServicePipeWireConfig.cmake.in`, `include/PhosphorServicePipeWire/`, `src/`, `tests/`. Imperative QML registration via `src/qmlregistration.cpp` (`qmlRegisterUncreatableType` for `PipeWireConnection` / `PwNode` / `PwNodeModel`, `qmlRegisterType` for `PwSinkModel` / `PwSourceModel` / `PwStreamModel`, `qmlRegisterSingletonType` for `PipeWireHost`), idempotency-guarded with `std::call_once`. This matches the sibling `phosphor-service-mpris` / `-upower` / `-sni` / `-icontheme` pattern: the lib ships no `.qml` files and installs nothing under `${KDE_INSTALL_QMLDIR}`; only the C++ shared library is installed and the QML module is process-global once `registerQmlTypes()` is called. The shell composition root (`src/shell/main.cpp`) calls `PhosphorServicePipeWire::registerQmlTypes()` alongside the four 2.0 sibling services so the QML module is process-global before any consumer mounts a `Phosphor.Service.PipeWire` import.
-2. **PipeWire core-connect lifecycle (M, ~3 days).** `PipeWireConnection` owns the `pw_main_loop` + `pw_context` + `pw_core`. Off-thread loop on a `QThread` (PipeWire's loop is not Qt-event-loop friendly), with a `QObject` bridge that posts events to the GUI thread via queued signals. Surface `pw_core` errors via the `error(QString)` signal + `connected → false` flip; the lib does not auto-reconnect (the shell drives backoff via `PipeWireConnection::connect()` / `PipeWireHost::reconnect()`). No public surface yet, just the lifecycle and the cross-thread plumbing.
+2. **PipeWire core-connect lifecycle (M, ~3 days).** `PipeWireConnection` owns the `pw_main_loop` + `pw_context` + `pw_core`. Off-thread loop on a `QThread` (PipeWire's loop is not Qt-event-loop friendly), with a `QObject` bridge that posts events to the GUI thread via queued signals. Surface `pw_core` errors via the `error(QString)` signal + `connected → false` flip; the lib does not auto-reconnect (the shell drives backoff via `PipeWireConnection::connectToDaemon()` / `PipeWireHost::reconnect()`). No public surface yet, just the lifecycle and the cross-thread plumbing.
 3. **Object registry: nodes (M-L, ~5 days).** The registry walk is folded into `PipeWireConnection::Private` rather than a standalone `PipeWireRegistryObserver` class. The Private holds the `pw_registry` listener and materialises one typed Qt object family: `PwNode` (parented to the connection, vended to consumers via `nodeAdded` / `nodeRemoved` signals). Property updates route through queued signals from the loop thread to the GUI thread. Filtering happens consumer-side via `PwNodeModel::mediaClasses` (`Audio/Sink`, `Audio/Source`, `Stream/Output/Audio`, `Stream/Input/Audio`); the connection surfaces every audio-class node and lets the models decide who they show. Ports, links, and devices are deliberately out of scope for 2.1 (the mixer use case is node-only); if a later phase needs them they land as sibling typed objects without changing the existing surface.
 4. **Sink/source/stream models (M, ~4 days).** `PwSinkModel`, `PwSourceModel`, `PwStreamModel` (per-app streams) as `QAbstractListModel` over the connection's filtered view. Role list pinned to: `node` (the `PwNode*` itself), `id`, `name`, `nick`, `description`, `mediaClass`, `channelCount`, `volumes` (per-channel linear amplitudes), `muted`, plus `Qt::DisplayRole` (`display`) that falls back through nick → description → name. Default-node membership is read from `PipeWireHost.defaultSinkName` / `defaultSourceName` rather than carried as a per-row role, so bindings stay reactive without the model re-emitting `dataChanged` on every default swap. Role enum integer values + role-name strings pinned via smoke tests (the Phase 2.0 sni-shell pattern).
 5. **Volume + mute write path (M, ~3 days).** `PwNode::setVolume(qreal)`, `setVolumes(QList<qreal>)`, `setMuted(bool)`. PipeWire writes go through `pw_node_set_param(SPA_PARAM_Props)` with a `SPA_PROP_channelVolumes` array; round-trip through the registry signal echo so the model reflects clamped values. Per-channel volume + Linear vs Cubic mapping decision (see Unknowns).
