+++ libs/phosphor-service-pipewire/src/pipewirehost.cpp
@@ -36,7 +36,21 @@ PipeWireHost::PipeWireHost(QObject* parent)
     d->connection->connectToDaemon();
 }
 
-PipeWireHost::~PipeWireHost() = default;
+PipeWireHost::~PipeWireHost()
+{
+    // Symmetric with the ctor's connectToDaemon() call: explicitly
+    // tear down the daemon connection before Private (and the
+    // PipeWireConnection it owns) unwinds. Letting `= default` rely on
+    // ~PipeWireConnection to do the disconnect implicitly mixes
+    // destructor-order quirks (PipeWire loop thread teardown vs the
+    // QObject signal table) with reasoning about exit paths;
+    // disconnecting here keeps the lifecycle visibly paired with the
+    // ctor's connectToDaemon() and means signal observers see a clean
+    // disconnect-then-destroy sequence on shutdown.
+    if (d->connection) {
+        d->connection->disconnectFromDaemon();
+    }
+}
 
 PipeWireConnection* PipeWireHost::connection() const
 {
